Change cast(double as json) to pad 0 and use scientific notation (#8498)
Summary:
Make the returned json padded with .0, if the input double or real
value equals to an integer. For example, cast(1.0 as json)
returns json '1' before and will return json '1.0' after this change.

Also if the magnitude of the input double or real value is greater
than or equal to 10^7, or less than 10^-3, return scientific
notation instead of standard notation. For example,
cast(0.0001 as json) returns json '0.0001' before and will return
json '1.0E-4' after this change.

This change also applies to cast(map|array|row as json) if the input
map, array or row has nested double or real elements.

It is then exactly the same behavior as cast(double as varchar).
And thus they will use the same util::Converter util.

Put this change behind the config legacy_cast = false, similar to
the change to cast(double as varchar).

It will also propagate to the return value format of
cast(cast(double as json) as varchar), and
json_format(cast(double as json)).

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8498

Reviewed By: kagamiori

Differential Revision: D53036540

Pulled By: gggrace14

fbshipit-source-id: 108250d8bea82ee22600a67296a5dcc56affa439